sensors-server/requirements.txt
ptvsd
gpiozero
Adafruit_DHT
paho-mqtt
Pillow
